+++ whitelist.py
@@ -1,3 +1,8 @@
 Color  # unused class (src/verna/verna.py:10)
 _.from_name  # unused method (src/verna/verna.py:32)
-_.from_rgba  # unused method (src/verna/verna.py:42)
+_.replace  # unused method (src/verna/verna.py:53)
+_.rgba  # unused method (src/verna/verna.py:92)
+_.to_float  # unused method (src/verna/verna.py:127)
+_.to_percentage  # unused method (src/verna/verna.py:139)
+_.to_int  # unused method (src/verna/verna.py:163)
+