+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+<h4> v0.1.0 () </h4>
+
+ - Allow assigning float values (ranging from 0.0 to 1.0) for red,
+   green and blue properties.
+ - Add `to_int()`, `to_float()` and `to_percentage()` functions.
+
 <h4> v0.0.2 (16 Oct 2020) </h4>
-                                                                                                                                      
- - Fix error in `Color.from_name()` so that it returns `Color` instead of `int`                                                       
- - Use setuptools for packaging  
+
+ - Fix error in `Color.from_name()` so that it returns `Color` instead of `int`.
+ - Use setuptools for packaging.
 
 <h4> v0.0.1 (15 Oct 2020) </h4>